requirements.txt
kmodes==0.11.1
matplotlib==3.5.1
numpy==1.22.3
pandas==1.4.3
psycopg2-binary
pymysql==1.0.2
PyYAML==3.13
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
tqdm==4.64.0
tslearn==0.5.2

